pyproject.toml
[tool.poetry]
authors = ["Calvin Laurenson <89622328+calvin-laurenson@users.noreply.github.com>"]
description = ""
name = "grosbeak"
version = "0"

[tool.poetry.dependencies]
Pillow = "9.1.0"
fastapi = "0.75.0"
loguru = "0.6.0"
pydantic = "1.9.0"
pymongo = {git = "https://github.com/mongodb/mongo-python-driver.git", rev = "fd512d5c90220a37341aca068bc9e3f969c8eead", extras = ["srv"]}
python = "^3.8"
python-multipart = "0.0.5"
uvicorn = {version = "0.17.6", extras = ["standard"]}

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
mypy = "^0.942"
types-Pillow = "^9.0.11"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]

[tool.poetry-dynamic-versioning]
enable = true
pattern = "^(?P<base>\\d+(\\.\\d+)*)"
style = "semver"
vcs = "git"

[tool.black]
src = "grosbeak/"
include = '\.py$'

[tool.mypy]
files = "./grosbeak/**/*.py"
namespace_packages = true
explicit_package_bases = true

[tool.mypy-pymongo]
ignore_missing_imports = true